demo using authentication vs authorizatison

+++ app/Http/Controllers/UserController.php
@@ -7,8 +7,10 @@ use App\Http\Services\GroupService;
 use App\Http\Services\UserService;
 use App\Role;
 use App\User;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Hash;
 
 class UserController extends Controller
 {
@@ -30,6 +32,9 @@ class UserController extends Controller
 
     public function showFormEdit($id)
     {
+        if (!Gate::allows('crud-user')){
+            abort(403);
+        }
         $user = $this->userService->getById($id);
         $roles = Role::all();
         $groups = $this->groupService->getAll();
@@ -39,6 +44,10 @@ class UserController extends Controller
 
     public function create()
     {
+        if (!Gate::allows('crud-user')){
+            abort(403);
+        }
+
         $groups = $this->groupService->getAll();
         $roles = Role::all();
         return view('layout.users.create', compact('groups', 'roles'));
@@ -46,11 +55,15 @@ class UserController extends Controller
 
     public function store(CreateUserRequest $request)
     {
+        if (!Gate::allows('crud-user')){
+            abort(403);
+        }
 
         DB::beginTransaction();
         try {
             $user = new User();
             $user->fill($request->all());
+            $user->password = Hash::make($request->password);
             $user->save();
             $user->roles()->sync($request->role);
             DB::commit();
@@ -62,6 +75,10 @@ class UserController extends Controller
     }
 
     public function delete($id) {
+        if (!Gate::allows('crud-user')){
+            abort(403);
+        }
+
         $user = $this->userService->getById($id);
         $user->delete();
         $result = [
